[Add] unit buy system
유닛 구매 시스템 초기 제작 완료

+++ Assets/Scripts/UnitInfo.cs
@@ -12,7 +12,7 @@ public class UnitInfo :ScriptableObject
     public string unitDescription;
     public int unitNumber;
     public bool isMine;
-    public float unitPrice;
+    public int unitPrice;
 
 
 }
